requirements.txt
Django
djangorestframework
markdown
django-filter
fpdf